Update ArenaManager.java
No need to loop through the arenas 3 times to check if the arena name starts with or ends with the given String. s.contains(sName) is all you need because it would return true if the arena name had the given substring in the beginning or end. No big deal really, I just thought it was a bit redundant.